# trading-service/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Trading Service API",
        "title": "Trading Service",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/actuaries": {
            "get": {
                "description": "Vraća listu svih zapisa aktuara iz baze podataka.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Actuaries"
                ],
                "summary": "Dobavljanje svih aktuara",
                "responses": {
                    "200": {
                        "description": "Uspešno dobavljeni svi aktuari",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/types.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/types.Actuary"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Greška u bazi",
                        "schema": {
                            "$ref": "#/definitions/types.Response"
                        }
                    }
                }
            },
            "post": {
                "description": "Kreira novi aktuar na osnovu dostavljenih podataka.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Actuaries"
                ],
                "summary": "Kreiranje novog aktuara",
                "parameters": [
                    {
                        "description": "Podaci za kreiranje aktuara",
                        "name": "actuary",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.ActuaryDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Uspešno kreiran aktuar",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/types.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/types.Actuary"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Neispravan format unosa ili poslati podaci nisu validni.",
                        "schema": {
                            "$ref": "#/definitions/types.Response"
                        }
                    },
                    "500": {
                        "description": "Greska u bazi.",
                        "schema": {
                            "$ref": "#/definitions/types.Response"
                        }
                    }
                }
            }
        },
        "/actuaries/filter": {
            "get": {
                "description": "Vraća listu aktuara filtriranu po imenu, prezimenu, email-u i/ili poziciji..",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Actuaries"
                ],
                "summary": "Filtriranje aktuara",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter po imenu (case-insensitive, partial match)",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter po prezimenu (case-insensitive, partial match)",
                        "name": "surname",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter po email-u (case-insensitive, partial match)",
                        "name": "email",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter po poziciji (exact match from user-service)",
                        "name": "position",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Uspešno filtrirani aktuari",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/types.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/dto.FilteredActuaryDTO"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Greska pri preuzimanju aktuara.",
                        "schema": {
                            "$ref": "#/definitions/types.Response"
                        }
                    }
                }
            }
        },
        "/actuaries/{ID}": {
            "get": {
                "description": "Vraća detalje jednog aktuara na osnovu njegovog ID-ja.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Actuaries"
                ],
                "summary": "Dobavljanje aktuara po ID-ju",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID aktuara",
                        "name": "ID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Uspešno dobavljen aktuar",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/types.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/types.Actuary"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Aktuar nije pronadjen",
                        "schema": {
                            "$ref": "#/definitions/types.Response"
                        }
                    },
                    "500": {
                        "description": "Greška u bazi",
                        "schema": {
                            "$ref": "#/definitions/types.Response"
                        }
                    }
                }
            }
        },
        "/actuaries/{id}": {
            "put": {
                "description": "Ažurira iznos limita za određeni aktuar prema ID-ju.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Actuaries"
                ],
                "summary": "Izmena limita aktuara",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID aktuara",
                        "name": "ID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Podaci za ažuriranje limita",
                        "name": "updateData",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateActuaryDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Uspešno ažurirani limiti aktuara",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/types.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/types.Actuary"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Neispravan format podataka",
                        "schema": {
                            "$ref": "#/definitions/types.Response"
                        }
                    },
                    "404": {
                        "description": "Aktuar nije pronadjen",
                        "schema": {
                            "$ref": "#/definitions/types.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.ActuaryDTO": {
            "type": "object",
            "required": [
                "department",
                "userID"
            ],
            "properties": {
                "department": {
                    "type": "string",
                    "enum": [
                        "supervisor",
                        "agent"
                    ]
                },
                "limitAmount": {
                    "type": "number"
                },
                "needApproval": {
                    "type": "boolean"
                },
                "usedLimit": {
                    "type": "number"
                },
                "userID": {
                    "type": "integer"
                }
            }
        },
        "dto.FilteredActuaryDTO": {
            "type": "object",
            "properties": {
                "department": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "firstName": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "lastName": {
                    "type": "string"
                },
                "limitAmount": {
                    "type": "number"
                },
                "needApproval": {
                    "type": "boolean"
                },
                "position": {
                    "type": "string"
                },
                "usedLimit": {
                    "type": "number"
                }
            }
        },
...
        "dto.UpdateActuaryDTO": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "string"
                },
                "reset": {
                    "type": "boolean"
                }
            }
        },
        "types.Actuary": {
            "type": "object",
            "properties": {
                "department": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "fullName": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "limit": {
                    "description": "Samo za agente",
                    "type": "number"
                },
                "needApproval": {
                    "description": "Da li orderi agenta trebaju supervizorsko odobrenje",
                    "type": "boolean"
                },
                "usedLimit": {
                    "description": "Samo za agente, resetuje se dnevno",
                    "type": "number"
                },
                "userId": {
                    "type": "integer"
                }
            }
        },
...
        "types.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Unesite token. Primer: \"Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}